requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-crispy-forms==1.13.0
django-model-utils==4.1.1
docker==5.0.3
idna==3.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
websocket-client==1.2.1
